added window toolbar & supported minimize/maximise window
