+++ BankVoiceWindow.xaml
@@ -1,4 +1,4 @@
-﻿<Window x:Class="dotnetAnima.BankVoiceWindow"
+﻿<Page x:Class="dotnetAnima.BankVoiceWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
@@ -6,13 +6,9 @@
         xmlns:local="clr-namespace:dotnetAnima"
         mc:Ignorable="d"
         Title="Bank Your Voice" 
-        Height="600" Width="920"
-        WindowStyle="None"
-        ResizeMode="NoResize"
-        Background="Transparent"
-        AllowsTransparency="True">
+        Height="600" Width="920">
 
-    <Border Background="Beige" CornerRadius="20">
+    <Border Background="Beige">
         <Grid>
             <Grid.ColumnDefinitions>
                 <ColumnDefinition Width="150"/>
@@ -195,5 +191,5 @@
 
         </Grid>
     </Border>
-</Window>
+</Page>
