+++ MainWindow.xaml
@@ -3,95 +3,15 @@
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:dotnetAnima"
         mc:Ignorable="d"
-        Title="Anima v3" 
-        Height="600" Width="920"
-        WindowStyle="None"
-        ResizeMode="NoResize"
-        Background="Transparent"
-        AllowsTransparency="True">
-
-    <Border Background="Beige" CornerRadius="20">
+        Title="Anima V3" 
+        Height="640" Width="920"
+        Background="Beige">
+    <Viewbox>
         <Grid>
-            <Grid.ColumnDefinitions>
-                <ColumnDefinition Width="150"/>
-                <ColumnDefinition Width="750"/>
-            </Grid.ColumnDefinitions>
-
-            <Grid.RowDefinitions>
-                <RowDefinition Height="100"/>
-                <RowDefinition Height="100"/>
-                <RowDefinition Height="420"/>
-            </Grid.RowDefinitions>
-
-            <!-- Anima Logo -->
-            <TextBlock Text="Anima v3"
-                       VerticalAlignment="Top"
-                       HorizontalAlignment="Center"
-                       Foreground="DarkBlue"
-                       Grid.Row="0"
-                       Grid.Column="0"
-                       FontSize="25"
-                       Margin="0,15,0,0"
-                       FontWeight="Bold">
-                <TextBlock.Effect>
-                    <DropShadowEffect Color="Black"
-                                      Direction="270"
-                                      ShadowDepth="5"
-                                      Opacity="0.5"/>
-                </TextBlock.Effect>
-            </TextBlock>
-
-            <!-- Start Button -->
-            <Button Content="START VOICE BANKING"
-             Name="startButton"
-             Foreground="White"
-             Style="{StaticResource MenuButtonTheme}" 
-             Grid.Column="1" 
-             Grid.Row="2" 
-             Margin="136,334,282,32"
-             Click="ButtonClick"/>
-
-            <!-- Anima Title Photo -->
-            <Image Source="Images/Anima Title.png" Grid.Row="0" Grid.RowSpan="2" Grid.ColumnSpan="2" Margin="3,0,-21,15"/>
-
-            <!-- Soundwave Photo -->
-            <Image Source="Images/soundwave.png" Grid.RowSpan="2" Margin="12,26,151,86" Grid.Row="1" Grid.Column="1"/>
-
-            <!-- Moto -->
-            <TextBlock Text="The comprehensive voice banking solution"
-            VerticalAlignment="Top"
-            HorizontalAlignment="Left"
-            Foreground="Black"
-            FontSize="18"
-            Margin="128,92,0,0"
-            FontWeight="Bold" Grid.Column="1" Grid.Row="1" RenderTransformOrigin="0.43,-1.284" Grid.RowSpan="2">
-                <TextBlock.Effect>
-                    <DropShadowEffect Color="Black"
-                       Direction="270"
-                       ShadowDepth="5"
-                       Opacity="0.3"/>
-                </TextBlock.Effect>
-            </TextBlock>
-
-            <!-- Description -->
-            <TextBlock
-            VerticalAlignment="Top"
-            HorizontalAlignment="Left"
-            Foreground="Black"
-            FontSize="18"
-            Name="desc"
-            FontStyle="Italic"
-            FontWeight="Medium" 
-            Grid.Row="2" 
-            Grid.ColumnSpan="2" 
-            Margin="130,257,0,0">
-            <Run Text="Voice banking involves recording your own voice, this can then be used to create a "/>
-            <LineBreak/>
-            <Run Language="en-gb" Text="&#x9;&#x9;    "/>
-            <Run Text="digital voice that can read text for you"/>
-            </TextBlock>
-
+            <Frame Name="MainFrame" NavigationUIVisibility="Hidden"/>
         </Grid>
-    </Border>
+    </Viewbox>
+    
 </Window>
