+++ MainWindow.xaml.cs
@@ -22,29 +22,11 @@ namespace dotnetAnima
     {
         public MainWindow()
         {
-            string path = @"../../animaProfiles";
+
             InitializeComponent();
-            if (Directory.Exists(path))
-            {
-                string[] directoriesWithinPath = Directory.GetFiles(path);
-                if (directoriesWithinPath.Length > 0)
-                {
-                    desc.Inlines.Clear();
-                    desc.Inlines.Add(new Run("                                                    Welcome Back!") { FontWeight = FontWeights.Bold });
-                    startButton.Content = "Text-to-Speech";
-                    startButton.Margin = new Thickness(136,319,282,47);
-                }
-            }
-
+            MainFrame.Navigate(new AnimaHomePage());
         }
 
-        private void ButtonClick(object sender, RoutedEventArgs e)
-        {
-            BankVoiceWindow bankWindow = new BankVoiceWindow();
-            this.Close();
-            bankWindow.Left = this.Left;
-            bankWindow.Top = this.Top;
-            bankWindow.Show();
-        }
+        
     }
 }
