+++ ManageVoicesWindow.xaml
@@ -1,4 +1,4 @@
-﻿<Window x:Class="dotnetAnima.ManageVoicesWindow"
+﻿<Page x:Class="dotnetAnima.ManageVoicesWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
@@ -6,12 +6,8 @@
         xmlns:local="clr-namespace:dotnetAnima"
         mc:Ignorable="d"
         Title="Manage Voices" 
-        Height="600" Width="920"
-        WindowStyle="None"
-        ResizeMode="NoResize"
-        Background="Transparent"
-        AllowsTransparency="True">
-    <Border Background="Beige" CornerRadius="20">
+        Height="600" Width="920">
+    <Border Background="Beige">
         <Grid>
             <Grid.ColumnDefinitions>
                 <ColumnDefinition Width="150"/>
@@ -47,19 +43,23 @@
                           Grid.Row="1"
                           Grid.RowSpan="2"
                           Grid.Column="0"
-                          Grid.ColumnSpan="2" Margin="0,0,245,33">
+                          Grid.ColumnSpan="2" Margin="0,7,245,26">
                 <Border Background="DeepSkyBlue" CornerRadius="20" Margin="20">
                     <StackPanel Margin="20,20,0,0"
                                 HorizontalAlignment="Center"
                                 VerticalAlignment="Top">
 
                         <!-- Your Voice -->
+
+
+                        <!-- Make this dynamic content -->
+                        <!-- Your Voice -->
                         <TextBlock Foreground="AntiqueWhite"
                                    FontSize="40"
                                    FontStyle="Normal"
                                    FontWeight="Bold"
                                    Text="Your Voice"
-                                   Margin="0,0,0,10"/>
+                                   Height="56" Width="587"/>
 
 
                         <!-- Make this dynamic content -->
@@ -78,7 +78,7 @@
                                    FontStyle="Normal"
                                    FontWeight="Bold"
                                    Text="Other Voices"
-                                   Margin="0,0,0,10"/>
+                                   Margin="0,0,0,10" Height="56"/>
 
                         <!-- Again dynamic -->
                         <CheckBox Content="Option 2"
@@ -162,4 +162,4 @@
         </Grid>
     </Border>
 
-</Window>
+</Page>
