+++ TextToSpeechWindow.xaml
@@ -1,4 +1,4 @@
-﻿<Window x:Class="dotnetAnima.TextToSpeechWindow"
+﻿<Page x:Class="dotnetAnima.TextToSpeechWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
@@ -6,12 +6,8 @@
         xmlns:local="clr-namespace:dotnetAnima"
         mc:Ignorable="d"
         Title="Text to Speech"
-        Height="600" Width="920"
-        WindowStyle="None"
-        Background="Transparent"
-        ResizeMode="NoResize"
-        AllowsTransparency="True">
-    <Border Background="Beige" CornerRadius="20">
+        Height="600" Width="920">
+    <Border Background="Beige">
         <Grid>
             <Grid.ColumnDefinitions>
                 <ColumnDefinition Width="150"/>
@@ -124,4 +120,4 @@
         </Grid>
     </Border>
     
-</Window>
+</Page>
