+++ dotnetAnima.csproj
@@ -106,6 +106,9 @@
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </ApplicationDefinition>
+    <Compile Include="AnimaHomePage.xaml.cs">
+      <DependentUpon>AnimaHomePage.xaml</DependentUpon>
+    </Compile>
     <Compile Include="BankVoiceWindow.xaml.cs">
       <DependentUpon>BankVoiceWindow.xaml</DependentUpon>
     </Compile>
@@ -123,6 +126,10 @@
     <Compile Include="TextToSpeechWindow.xaml.cs">
       <DependentUpon>TextToSpeechWindow.xaml</DependentUpon>
     </Compile>
+    <Page Include="AnimaHomePage.xaml">
+      <SubType>Designer</SubType>
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
     <Page Include="BankVoiceWindow.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
